+++ src/generated_load.cpp
@@ -24,6 +24,7 @@
 #include "data_server.hpp"
 #include "readable_id.hpp"
 #include <optional>
+#include "dnpactivities.hpp"
 
 namespace {
 
@@ -66,6 +67,14 @@ std::string store_relation(base_object &generated_object,
         }
     }
     return uuid;
+}
+
+// Helper to determine whether the setting for DNP activities URL should
+// be read as it is (a URL) or as a local file path (in which case we want
+// to read it differently from the config file so relative paths work).
+bool looks_like_url(const std::string& str) {
+    auto upper = toUpper(str);
+    return startsWith(upper, "HTTP://") || startsWith(upper, "HTTPS://") || startsWith(upper, "FILE://");
 }
 
 /**
@@ -140,6 +149,45 @@ std::shared_ptr<object_list> get_generated_activity(const std::string &type,
         group_attribute = conf.get(group_match_attribute);
     }
 
+    // DNP Activities settings
+    std::shared_ptr<DNPActivities> dnpactivities;
+    const std::string test_activities_url_variable = type + "-national-test-activities-url";
+    std::string test_activities_url = conf.get(test_activities_url_variable, true);
+    if (!test_activities_url.empty()) {
+        if (!looks_like_url(test_activities_url)) {
+            // Assume it's a path instead, in which case we'll get it with get_path instead so
+            // relative paths are converted to absolute.
+            test_activities_url = conf.get_path(test_activities_url_variable, true);
+            // Then convert it to a file:// URL
+#ifdef _WIN32
+            std::string separator = "/";
+#else
+            std::string separator = "";
+#endif
+            test_activities_url = "file://" + separator + test_activities_url;
+        }
+        // This will throw and terminate the load process if it fails
+        dnpactivities = create_activities_from_url(test_activities_url);
+    }
+    // Which attribute (in the StudentGroup) contains the test name
+    std::string activity_name_attribute = conf.get(type + "-national-test-activity-name-attribute", true);
+    // Which attribute (in the generated Activity object) should be used to store the UUID of the test activity
+    std::string activity_id_attribute = conf.get(type + "-national-test-activity-id-attribute", true);
+    if (activity_id_attribute == "") {
+        activity_id_attribute = "parentActivity";
+    }
+    // Converting names to ids will be done if dnpactivities have been loaded and the 
+    // configuration has specified where to find the names.
+    const bool convert_activity_names_to_ids = dnpactivities != nullptr && !activity_name_attribute.empty();
+
+    // Settings for deducing the test activity name suffix
+    const auto deduce_from_school_type_attribute = conf.get(type + "-deduce-test-activity-suffix-from-school-type-attribute", true);
+    const auto deduce_from_student_id_attribute = conf.get(type + "-deduce-test-activity-suffix-from-members-with-school-year", true);
+    const auto deduce_from_school_year_attribute = conf.get(type + "-deduce-test-activity-suffix-from-school-year-attribute", true);
+
+    // Deducing test activity name suffixes will be done if attributes have been configured for schoolYear and schoolType
+    const bool deduce_suffixes = !deduce_from_school_type_attribute.empty() && !deduce_from_school_year_attribute.empty();
+
     data_server &server = data_server::instance();
     auto student_groups = server.get_by_type(master_type);
     auto employments = server.get_by_type(related_type.first);
@@ -173,6 +221,45 @@ std::shared_ptr<object_list> get_generated_activity(const std::string &type,
             if (!found_proper_employment && !employments.empty()) {
                 auto employment = employments.front();
                 generated_object.append_values(pair_to_string(related_type), {employment->get_uid()});
+            }
+        }
+
+        if (convert_activity_names_to_ids) {
+            auto values = student_group.second->get_values(activity_name_attribute);
+            // Multi-valued is a bit weird but we'll handle it...
+            std::vector<std::string> ids;
+            for (const std::string &value : values) {
+                auto id = dnpactivities->name_to_id(value);
+                if (id) {
+                    ids.push_back(id.get());
+                } else if (deduce_suffixes) {
+                    std::shared_ptr<object_vector> students;
+                    if (!deduce_from_student_id_attribute.empty()) {
+                        students = std::make_shared<std::vector<std::shared_ptr<base_object>>>();
+                        auto student_ids = student_group.second->get_values(deduce_from_student_id_attribute);
+                        auto student_type = string_to_pair(deduce_from_student_id_attribute).first;
+
+                        for (auto &student_id : student_ids) {
+                            auto student = data_server::instance().get_by_id(student_type, student_id);
+                            if (student != nullptr) {
+                                students->push_back(student);
+                            }
+                        }
+                    }
+
+                    // See if adding a deduced suffix turns the name into a valid activity name
+                    std::string with_suffix = value + deduce_suffix(student_group.second, 
+                                                                    deduce_from_school_type_attribute, 
+                                                                    deduce_from_school_year_attribute,
+                                                                    students);
+                    id = dnpactivities->name_to_id(with_suffix);
+                    if (id) {
+                        ids.push_back(id.get());
+                    }
+                }
+            }
+            if (!ids.empty()) {
+                generated_object.add_attribute(activity_id_attribute, ids);
             }
         }
 
@@ -501,4 +588,45 @@ std::shared_ptr<object_list> get_generated(const std::string &type,
     if (list && !list->empty())
         load_related(type, list, load_logger);
     return list;
+}
+
+std::string deduce_suffix(std::shared_ptr<base_object> student_group, 
+    const std::string& deduce_from_school_type_attribute, 
+    const std::string& deduce_from_school_year_attribute,
+    std::shared_ptr<object_vector> members_with_school_year) {
+    auto school_types = student_group->get_values(deduce_from_school_type_attribute);
+    if (school_types.size() != 1) {
+        return "";
+    }
+    std::string school_type = school_types[0];
+
+    if (school_type != "GR") {
+        if (school_type == "VUX") {
+            school_type = "VUXGY";
+        }
+        return "_" + school_type;
+    } else {
+        if (members_with_school_year != nullptr) {
+            std::vector<std::string> school_years;
+            for (const auto &member : *members_with_school_year) {
+                auto students_school_years = member->get_values(deduce_from_school_year_attribute);
+                if (!students_school_years.empty()) {
+                    school_years.push_back(students_school_years.front());
+                }
+            }
+
+            if (school_years.empty()) {
+                return "";
+            } else {
+                return "_" + most_common<std::string>(school_years.begin(), school_years.end());
+            }
+        } else {
+            auto school_years = student_group->get_values(deduce_from_school_year_attribute);
+            if (school_years.size() != 1) {
+                return "";
+            } else {
+                return "_" + school_years.front();
+            }
+        }
+    }
 }